OPEN-13 refactor ChannelUIController
